fix(deps): update to require tf >=1.9
